feat(artworks): persist and restore selected page using query params and sessionStorage
- Move pagination logic to a client-only component wrapped in Suspense
- Fix issue where returning from detail page reset to page 1
- Ensure compatibility with Vercel static export constraints
